Janitor: Remove unecessary cast
Prefer a u8 literal over an i32 literal that is then cast to a u8

+++ sixtyfps_runtime/rendering_backends/gl/graphics_window.rs
@@ -301,7 +301,7 @@ impl GraphicsWindow {
             {
                 window_item.background()
             } else {
-                RgbaColor { red: 255 as u8, green: 255, blue: 255, alpha: 255 }.into()
+                RgbaColor { red: 255_u8, green: 255, blue: 255, alpha: 255 }.into()
             };
 
             let mut renderer = window.backend.borrow_mut().new_renderer(
